requirements-pi.txt
click==6.7
python-socketio==4.0.1
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
picamera==1.13
Werkzeug==0.12.2
